requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
Django==3.1.6
googlesearch-python==2020.0.2
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.20.1
oauthlib==3.1.0
pandas==1.2.2
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
sklearn==0.0
socialreaper==0.3.0
soupsieve==2.0.1
sqlparse==0.4.1
threadpoolctl==2.1.0
urllib3==1.25.9
